On ``setRotation`` we should aways wright the variables as integers, even when the variable is a float.
Explanation, in portuguese:

Assim, um número que tem uma casa decimal, por exemplo 10.8 , seria um write (parametro, 108). Se a variável fosse pore xemplo 32.75, seria um write (parametro, 3275). No caso do cskfan, self.sk.write_parm('01.21',float(freq)) precisa ser self.sk.write_parm('01.21',freq) e se a frequencia desejada é por exemplo 10.6 Hz, o valor de freq deve ser 106 e a gente descobre o número de casas decimais olhando o default no manual ou no CTSoft.

+++ chimera_commandersk/instruments/cskfan.py
@@ -57,7 +57,7 @@ class CSKFan(ChimeraObject):
         @param freq: Frequency in Hz
         @type  freq: float
         """
-        self.sk.write_parm('01.21',float(freq))
+        self.sk.write_parm('01.21',int(freq))
 
     def getDirection(self):
         return self.direction
